Move logic from MainFragment to ClientFragment
