+++ client/app/src/main/java/engineer/mathsoftware/ep/tcpfs/ClientFragment.kt
@@ -4,13 +4,18 @@
 
 package engineer.mathsoftware.ep.tcpfs
 
+import android.app.Activity
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.navigation.fragment.findNavController
+import androidx.documentfile.provider.DocumentFile
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.lifecycleScope
 import engineer.mathsoftware.ep.tcpfs.databinding.FragmentClientBinding
+import kotlinx.coroutines.launch
 
 class ClientFragment : Fragment() {
     private var _binding: FragmentClientBinding? = null
@@ -18,6 +23,7 @@ class ClientFragment : Fragment() {
     // This property is only valid between onCreateView and
     // onDestroyView.
     private val binding get() = _binding!!
+    private lateinit var client: Client
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -29,14 +35,75 @@ class ClientFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        binding.buttonSecond.setOnClickListener {
-            findNavController().navigate(R.id.action_SecondFragment_to_FirstFragment)
+        binding.buttonUpload.setOnClickListener {
+            chooseFileToUpload()
         }
+        connect()
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
+        disconnect()
+    }
+
+    override fun onActivityResult(
+        requestCode: Int,
+        resultCode: Int,
+        data: Intent?
+    ) {
+        super.onActivityResult(requestCode, resultCode, data)
+        if (resultCode != Activity.RESULT_OK) {
+            return
+        }
+        if (data == null) {
+            return
+        }
+        when (requestCode) {
+            PICKFILE_REQUEST_CODE -> readFileToUpload(data.data)
+        }
+    }
+
+    private fun connect() {
+        lifecycleScope.launch {
+            val c = Client.newInstance()
+
+            if (c == null) {
+                handleConnectionFailed()
+            }
+            else {
+                println("connected")
+                client = c
+            }
+        }
+    }
+
+    private fun disconnect() {
+        lifecycleScope.launch {
+            client.disconnect()
+        }
+    }
+
+    private fun chooseFileToUpload() {
+        val intent = Intent(Intent.ACTION_OPEN_DOCUMENT).apply {
+            addCategory(Intent.CATEGORY_OPENABLE)
+            type = "*/*"
+        }
+        startActivityForResult(intent, PICKFILE_REQUEST_CODE)
+    }
+
+    private fun readFileToUpload(data: Uri?) {
+        var bytes = ByteArray(0)
+        val file = data?.let {
+            DocumentFile.fromSingleUri(requireContext(), it)
+        }?.name.toString()
+        if (data != null) {
+            bytes = read(requireContext().contentResolver, data)
+        }
+
+        lifecycleScope.launch {
+            client.file = file
+            client.upload(bytes)
+        }
     }
 }
