+++ client/app/src/main/java/engineer/mathsoftware/ep/tcpfs/MainFragment.kt
@@ -4,19 +4,14 @@
 
 package engineer.mathsoftware.ep.tcpfs
 
-import android.app.Activity
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.documentfile.provider.DocumentFile
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.viewbinding.ViewBindings
 import com.google.android.material.snackbar.Snackbar
 import engineer.mathsoftware.ep.tcpfs.databinding.FragmentMainBinding
 import kotlinx.coroutines.launch
@@ -24,9 +19,6 @@ import org.json.JSONException
 
 const val PICKFILE_REQUEST_CODE = 1
 
-/**
- * A simple [Fragment] subclass as the default destination in the navigation.
- */
 class MainFragment : Fragment() {
     private val channels = ArrayList<String>()
     private var _binding: FragmentMainBinding? = null
@@ -47,12 +39,8 @@ class MainFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        channelsAdapter = ChannelsAdapter(channels) {subscribe(it)}
+        channelsAdapter = ChannelsAdapter(channels) { subscribe(it) }
         initChannelList()
-
-        // binding.buttonUpload.setOnClickListener {
-        //     chooseFileToUpload()
-        // }
         connect()
     }
 
@@ -67,23 +55,6 @@ class MainFragment : Fragment() {
         super.onDestroyView()
         _binding = null
         disconnect()
-    }
-
-    override fun onActivityResult(
-        requestCode: Int,
-        resultCode: Int,
-        data: Intent?
-    ) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (resultCode != Activity.RESULT_OK) {
-            return
-        }
-        if (data == null) {
-            return
-        }
-        when (requestCode) {
-            PICKFILE_REQUEST_CODE -> readFileToUpload(data.data)
-        }
     }
 
     private fun connect() {
@@ -108,36 +79,14 @@ class MainFragment : Fragment() {
         }
     }
 
-    private fun disconnect() {
-        lifecycleScope.launch {
-            client.disconnect()
-        }
-    }
-
     private fun subscribe(channel: String) {
+        disconnect()
         findNavController().navigate(R.id.action_FirstFragment_to_SecondFragment)
     }
 
-    private fun chooseFileToUpload() {
-        val intent = Intent(Intent.ACTION_OPEN_DOCUMENT).apply {
-            addCategory(Intent.CATEGORY_OPENABLE)
-            type = "*/*"
-        }
-        startActivityForResult(intent, PICKFILE_REQUEST_CODE)
-    }
-
-    private fun readFileToUpload(data: Uri?) {
-        var bytes = ByteArray(0)
-        val file = data?.let {
-            DocumentFile.fromSingleUri(requireContext(), it)
-        }?.name.toString()
-        if (data != null) {
-            bytes = read(requireContext().contentResolver, data)
-        }
-
+    private fun disconnect() {
         lifecycleScope.launch {
-            client.file = file
-            client.upload(bytes)
+            client.disconnect()
         }
     }
 
@@ -146,14 +95,14 @@ class MainFragment : Fragment() {
         channels.addAll(values)
         channelsAdapter.notifyDataSetChanged()
     }
+}
 
-    private fun handleConnectionFailed() {
-        Snackbar.make(
-            requireView(),
-            "Fail to connect",
-            Snackbar.LENGTH_LONG
-        )
-            .setAction("Action", null)
-            .show()
-    }
+fun Fragment.handleConnectionFailed() {
+    Snackbar.make(
+        requireView(),
+        "Fail to connect",
+        Snackbar.LENGTH_LONG
+    )
+        .setAction("Action", null)
+        .show()
 }
