+++ client/app/src/main/res/layout/fragment_main.xml
@@ -27,20 +27,4 @@
         android:id="@+id/channelList"
         android:layout_width="match_parent" android:layout_height="match_parent"
         />
-
-    <!--<LinearLayout-->
-    <!--    android:orientation="horizontal"-->
-    <!--    android:layout_width="match_parent"-->
-    <!--    android:layout_height="wrap_content"-->
-    <!--    tools:layout_editor_absoluteY="1dp"-->
-    <!--    tools:layout_editor_absoluteX="1dp"-->
-    <!--    >-->
-
-    <!--    <Button-->
-    <!--        android:id="@+id/button_upload"-->
-    <!--        android:layout_width="match_parent"-->
-    <!--        android:layout_height="wrap_content"-->
-    <!--        android:text="@string/upload"-->
-    <!--        />-->
-    <!--</LinearLayout>-->
 </LinearLayout>
